requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pdfminer.six
python-docx
mail-parser
langchain
langchain-google-genai
sentence-transformers
pinecone
pinecone[grpc]
boto3
pydantic-settings
python-multipart